[Refactor] ChatRoom Container/Presenter로 나눔
ChatRoom에 대한 데이터 로직과 스타일 로직의 분리
ChatRoom 컴포넌트가 커짐에 따라
