+++ client/src/composition/ChatRooms/ChatRoom/index.tsx
@@ -1,0 +1,2 @@
+import ChatRoomContainer from './ChatRoomContainer';
+export default ChatRoomContainer;
